feat: authentication and authorization

+++ src/modules/user/routes.ts
@@ -1,4 +1,4 @@
-import { FastifyInstance } from "fastify";
+import { FastifyInstance, FastifyRequest } from "fastify";
 import { errorSchema } from "../../utils/error-handling/errorSchema";
 import {
   CreateUserDTO,
@@ -10,13 +10,18 @@ import {
 } from "./dto/dto";
 import { UserDTOMapper } from "./dto/mapper";
 import { userServiceFactory } from "./factories";
+import { authServiceFactory } from '../auth/factories';
+import { User, UserRole } from './model';
+import { authFactory } from '../../utils/security/auth';
+import { ErrorCodes } from '../../utils/error-handling/errorCodes';
 
 const userService = userServiceFactory();
+const auth = authFactory(authServiceFactory());
 
 export async function userRoutes(app: FastifyInstance) {
   app.route({
     method: "POST",
-    url: "/users",
+    url: "/users/new-account",
     schema: {
       body: createUserSchema,
       response: {
@@ -30,7 +35,7 @@ export async function userRoutes(app: FastifyInstance) {
 
       const existingUser = await userService.findByEmail(data.email);
       if (existingUser) {
-        return reply.status(409).send({ error: "Email already in use" });
+        return reply.status(409).send({ error: ErrorCodes.Conflict, message: "Email already in use" });
       }
 
       const user = await userService.create(UserDTOMapper.toCreateModel(data));
@@ -48,7 +53,8 @@ export async function userRoutes(app: FastifyInstance) {
         200: userListSchema,
       },
     },
-    handler: async (_, reply) => {
+    preValidation: auth([UserRole.Admin]),
+    handler: async (request, reply) => {
       const users = await userService.findAll();
       const usersDTO = users.map((user) => UserDTOMapper.toDto(user));
       return reply.send(usersDTO);
@@ -62,15 +68,22 @@ export async function userRoutes(app: FastifyInstance) {
       params: idParamSchema,
       response: {
         200: userSchema,
-        404: errorSchema,
+        404: errorSchema
       },
     },
+    preValidation: auth([UserRole.Student, UserRole.Admin]),
     handler: async (request, reply) => {
       const { id } = request.params as { id: string };
+      const { authenticated } = request.body as { authenticated: User };
+
+      if (authenticated.id !== id && authenticated.role !== UserRole.Admin) {
+        return reply.status(404).send({ error: ErrorCodes.NotFound, message: "User not found" });
+      }
+
       const user = await userService.findById(id);
 
       if (!user) {
-        return reply.status(404).send({ error: "User not found" });
+        return reply.status(404).send({ error: ErrorCodes.NotFound, message: "User not found" });
       }
 
       const userDTO = UserDTOMapper.toDto(user);
@@ -89,13 +102,20 @@ export async function userRoutes(app: FastifyInstance) {
         404: errorSchema,
       },
     },
+    preValidation: auth([UserRole.Student, UserRole.Admin]),
     handler: async (request, reply) => {
       const { id } = request.params as { id: string };
+      const { authenticated } = request.body as { authenticated: User };
+
+      if (authenticated.id !== id && authenticated.role !== UserRole.Admin) {
+        return reply.status(404).send({ error: ErrorCodes.NotFound, message: "User not found" });
+      }
+
       const updateData = request.body as Partial<CreateUserDTO>;
 
       const user = await userService.update(id, updateData);
       if (!user) {
-        return reply.status(404).send({ error: "User not found" });
+        return reply.status(404).send({ error: ErrorCodes.NotFound, message: "User not found" });
       }
 
       const userDTO = UserDTOMapper.toDto(user);
@@ -113,12 +133,19 @@ export async function userRoutes(app: FastifyInstance) {
         404: errorSchema,
       },
     },
+    preValidation: auth([UserRole.Student, UserRole.Admin]),
     handler: async (request, reply) => {
       const { id } = request.params as { id: string };
+      const { authenticated } = request.body as { authenticated: User };
+
+      if (authenticated.id !== id && authenticated.role !== UserRole.Admin) {
+        return reply.status(404).send({ error: ErrorCodes.NotFound, message: "User not found" });
+      }
+
       const deleted = await userService.delete(id);
 
       if (!deleted) {
-        return reply.status(404).send({ error: "User not found" });
+        return reply.status(404).send({ error: ErrorCodes.NotFound, message: "User not found" });
       }
 
       return reply.status(204).send();
